+++ core/server/api_container/server/startosis_engine/startosis_errors/interpretation_error_test.go
@@ -1,6 +1,7 @@
 package startosis_errors
 
 import (
+	"errors"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
@@ -39,4 +40,15 @@ func TestInterpretationError_serializationFromStacktrace(t *testing.T) {
 	at [13:12]: <toplevel>
 	at [0:0]: add_service`
 	require.Equal(t, expectedOutput, errorToSerialize.Error())
+}
+
+func TestInterpretationError_WithCausedBy(t *testing.T) {
+	rootCause := errors.New("root cause error")
+	levelOneInterpretationError := WrapWithInterpretationError(rootCause, "This is the root interpretation error")
+	userVisibleInterpretationError := WrapWithInterpretationError(levelOneInterpretationError, "An error happened!")
+
+	expectedErrorMessage := `An error happened!
+	Caused by: This is the root interpretation error
+	Caused by: root cause error`
+	require.Equal(t, expectedErrorMessage, userVisibleInterpretationError.Error())
 }